reinstate modules
