+++ app/controllers/stash_api/curation_activity_controller.rb
@@ -1,6 +1,7 @@
 require_dependency 'api_application_controller'
-require_dependency 'api_datasets_controller'
+require_relative 'stash_api/datasets_controller'
 
+module StashApi
   class ApiCurationActivityController < ApiApplicationController
     before_action :require_json_headers
     before_action :doorkeeper_authorize!
@@ -56,7 +57,7 @@ require_dependency 'api_datasets_controller'
     end
 
     def initialize_stash_identifier(id)
-      ds = DatasetsController.new
+      ds = StashApi::DatasetsController.new
       @stash_identifier = ds.get_stash_identifier(id)
       render json: { error: "cannot find dataset with identifier #{id}" }.to_json, status: 404 if @stash_identifier.nil?
     end
@@ -127,4 +128,4 @@ require_dependency 'api_datasets_controller'
     end
 
   end
-
+end
