+++ app/controllers/stash_api/users_controller.rb
@@ -1,5 +1,6 @@
 require_dependency 'api_application_controller'
 
+module StashApi
   class UsersController < ApiApplicationController
 
     before_action :require_json_headers
@@ -52,4 +53,4 @@ require_dependency 'api_application_controller'
       }
     end
   end
-
+end
