+++ app/models/stash_api/dataset_parser/base_parser.rb
