+++ app/models/stash_api/dataset_parser/temporal_coverages.rb
