+++ app/models/stash_api/version/metadata.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-
+module StashApi
   class Version
     class Metadata
 
@@ -72,3 +72,4 @@
       end
     end
   end
+end
