+++ app/models/stash_api/version/metadata/authors.rb
@@ -2,6 +2,7 @@
 
 require_relative 'metadata_item'
 
+module StashApi
   class Version
     class Metadata
       class Authors < MetadataItem
@@ -21,3 +22,4 @@ require_relative 'metadata_item'
       end
     end
   end
+end
