+++ app/models/stash_api/version/metadata/funders.rb
@@ -2,6 +2,7 @@
 
 require_relative 'metadata_item'
 
+module StashApi
   class Version
     class Metadata
       class Funders < MetadataItem
@@ -19,3 +20,4 @@ require_relative 'metadata_item'
       end
     end
   end
+end
