+++ app/models/stash_api/version/metadata/keywords.rb
@@ -2,6 +2,7 @@
 
 require_relative 'metadata_item'
 
+module StashApi
   class Version
     class Metadata
       class Keywords < MetadataItem
@@ -12,3 +13,4 @@ require_relative 'metadata_item'
       end
     end
   end
+end
