+++ app/models/stash_api/version/metadata/metadata_item.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+module StashApi
   class Version
     class Metadata
       class MetadataItem
@@ -11,3 +12,4 @@
       end
     end
   end
+end
