+++ app/models/stash_api/version/metadata/temporal_coverages.rb
@@ -2,6 +2,7 @@
 
 require_relative 'metadata_item'
 
+module StashApi
   class Version
     class Metadata
       class TemporalCoverages < MetadataItem
@@ -15,3 +16,4 @@ require_relative 'metadata_item'
       end
     end
   end
+end
