+++ config/initializers/stash_metadata_engine.rb
@@ -1,1 +1,1 @@
-APP_CONFIG.metadata_engine.constantize.config_resource_patch
+APP_CONFIG.metadata_engine.constantize.config_author_patch
