+++ stash/stash_datacite/lib/stash_datacite.rb
@@ -1,10 +1,8 @@
 require 'stash_datacite/engine'
-require 'stash_datacite/resource_patch'
 module StashDatacite
 
-  def self.config_resource_patch
+  def self.config_author_patch
     Rails.application.config.to_prepare do
-      StashDatacite::ResourcePatch.associate_with_resource(StashEngine::Resource)
       # Authorpatch also has to be here, otherwise there were places it wasn't patching before use, resulting in errors
       StashDatacite::AuthorPatch.patch! unless StashEngine::Author.method_defined?(:affiliation)
     end
